requirements.txt
Flask==2.3.2
openai==0.27.8
elevenlabs==0.2.4
python-dotenv==1.0.0
gunicorn==20.1.0  # maybe for deploy